Remove adminhelpers class and move its functionality into the admin page
